Add -11 memory option
Uses 5.5GB of ram, lol. This is the limit since the hash table takes
32 bit hashes for 32 / 64 bit compat and is 4 GB in size.

+++ CM.hpp
@@ -175,8 +175,8 @@ public:
 	MatchModelType match_model;
 
 	// Hash table
-	uint32_t hash_mask;
-	uint32_t hash_alloc_size;
+	size_t hash_mask;
+	size_t hash_alloc_size;
 	MemMap hash_storage;
 	uint8_t *hash_table;
